pyproject.toml
[tool.poetry]
name = "bugzilla-dashboard"
version = "0.1.0"
description = ""
authors = ["Armen Zambrano G. <armenzg@mozilla.com>"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^2.7"
PyYAML = "^5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
